+++ controller/counter.go
@@ -30,28 +30,6 @@ func (e *Env) GetAllBatchCounter(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintf(w, string(output))
 }
 
-func (e *Env) AllBatchPrice(w http.ResponseWriter, r *http.Request) {
-	log.Println("call AllMachineBatchSale()")
-	w.Header().Set("Server", "nava Stock")
-	w.Header().Set("Content-Type", "application/json")
-	w.Header().Set("Access-Control-Allow-Origin", "*")
-
-	rs := api.Response{}
-	p := model.BatchPrice{}
-	prices, err := p.All(e.DB)
-	if err != nil {
-		rs.Status = api.ERROR
-		rs.Message = err.Error()
-		w.WriteHeader(http.StatusNoContent)
-	} else {
-		rs.Status = api.SUCCESS
-		rs.Data = prices
-		w.WriteHeader(http.StatusOK)
-	}
-	output, _ := json.Marshal(rs)
-	fmt.Fprintf(w, string(output))
-}
-
 func (e *Env) NewBatchCounter(w http.ResponseWriter, r *http.Request) {
 	log.Println("call AllMachineBatchSale()")
 	w.Header().Set("Server", "nava Stock")
